requirements.txt
torch>=2.0.0
numpy>=1.21.0
sounddevice>=0.4.5
transformers>=4.30.0
openai>=1.0.0
aiohttp>=3.8.0
python-dotenv>=0.19.0
scipy>=1.7.0
librosa>=0.9.0
pyttsx3>=2.90
requests>=2.26.0
beautifulsoup4>=4.9.3 